Move TestConstantsLoader test util
