+++ ethereum/spec/build.gradle
@@ -10,6 +10,7 @@ dependencies {
 
     testFixturesImplementation project(':infrastructure:unsigned')
     testFixturesImplementation testFixtures(project(':bls'))
+    testFixturesImplementation project(':infrastructure:io')
 
     testFixturesApi project(':pow')
     testFixturesApi 'com.google.guava:guava'
